create movies module
