+++ server/__init__.py
@@ -4,6 +4,7 @@ import os
 
 from server.auth import auth_bp
 from server.main import main_bp
+from server.movies import movies_bp
 
 from .commands import create_tables, create_users_movies
 from .extensions import db, guard, migrate
@@ -30,6 +31,7 @@ def create_app(config_file='settings.py'):
     # register routes
     app.register_blueprint(main_bp)
     app.register_blueprint(auth_bp)
+    app.register_blueprint(movies_bp)
 
     # add cli commands
     app.cli.add_command(create_tables)
